Add scores to level select Add timer to all levels player can write name when wins

+++ Assets/Scripts/Models/ScoreModel.cs
@@ -9,7 +9,7 @@ namespace Assets.Scripts.Models
         public int LevelNumber;
 
         public int DeathCount;
-        public int Time;
+        public float Time;
     }
 
     [Serializable]
